requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Deprecated==1.2.14
idna==3.4
jwt==1.3.1
pycparser==2.21
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.0.2
wrapt==1.15.0
